pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dataguard"
version = "0.0.1"
description = "Data Guard is a powerful tool to monitor data pipelines and ensure data quality."
authors = [
    { name = "Sumz" },
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache Software License (Apache 2.0)"}
keywords = [
    "data quality",
    "data engineering",
]
classifiers = [
    "Development Status :: 4 - Alpha",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
  "cuallee~=0.15.2",
  "python-ulid~=2.7.0"
]

[project.optional-dependencies]
dev = [
  "black==24.8.0",
  "ruff==0.6.4",
  "pre-commit==3.8.0",
]
pandas = [
  "cuallee[pandas]~=0.15.2"
]
pyspark = [
  "cuallee[pyspark]~=0.15.2",
  "delta-spark~=3.1.0"
]
snowpark = [
  "cuallee[snowpark]~=0.15.2",
]
test = [
  "pytest",
  "pytest-cov",
  "pendulum >= 2.1.2",
  "cuallee~=0.15.2",
]
all = [
  "dataguard[dev,pyspark,snowpark,pandas,test]"
]

[project.urls]
"Homepage" = "https://github.com/SumzCol/dataguard"
"Bug Tracker" = "https://github.com/SumzCol/dataguard"